requirements.txt
httpx==0.27.2
pandas==2.2.3
python-dotenv==1.0.1
schedule==1.2.2
tqdm==4.66.5
backoff==2.2.1
duckdb==1.1.3

